core: dedupe approval not-found handling
