+++ extensions/telegram/src/exec-approval-resolver.ts
@@ -1,6 +1,9 @@
 import type { OpenClawConfig } from "openclaw/plugin-sdk/config-runtime";
 import { createOperatorApprovalsGatewayClient } from "openclaw/plugin-sdk/gateway-runtime";
-import type { ExecApprovalReplyDecision } from "openclaw/plugin-sdk/infra-runtime";
+import {
+  isApprovalNotFoundError,
+  type ExecApprovalReplyDecision,
+} from "openclaw/plugin-sdk/infra-runtime";
 
 export type ResolveTelegramExecApprovalParams = {
   cfg: OpenClawConfig;
@@ -10,38 +13,6 @@ export type ResolveTelegramExecApprovalParams = {
   allowPluginFallback?: boolean;
   gatewayUrl?: string;
 };
-
-const INVALID_REQUEST = "INVALID_REQUEST";
-const APPROVAL_NOT_FOUND = "APPROVAL_NOT_FOUND";
-
-function readErrorCode(value: unknown): string | null {
-  return typeof value === "string" && value.trim() ? value : null;
-}
-
-function readApprovalNotFoundDetailsReason(value: unknown): string | null {
-  if (!value || typeof value !== "object" || Array.isArray(value)) {
-    return null;
-  }
-  const reason = (value as { reason?: unknown }).reason;
-  return typeof reason === "string" && reason.trim() ? reason : null;
-}
-
-function isApprovalNotFoundError(err: unknown): boolean {
-  if (!(err instanceof Error)) {
-    return false;
-  }
-  const gatewayCode = readErrorCode((err as { gatewayCode?: unknown }).gatewayCode);
-  if (gatewayCode === APPROVAL_NOT_FOUND) {
-    return true;
-  }
-
-  const detailsReason = readApprovalNotFoundDetailsReason((err as { details?: unknown }).details);
-  if (gatewayCode === INVALID_REQUEST && detailsReason === APPROVAL_NOT_FOUND) {
-    return true;
-  }
-
-  return /unknown or expired approval id/i.test(err.message);
-}
 
 export async function resolveTelegramExecApproval(
   params: ResolveTelegramExecApprovalParams,
